requirements.txt
kagglehub>=0.2.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
streamlit>=1.30.0
